requirements.txt
# WordPress REST API Publisher - Requirements
# Python 3.8+

# Shared core package (keeps all client repos aligned)
-e ./packages/core_py

# HTTP requests to WordPress API
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Image processing (optional, for optimization)
Pillow>=10.0.0

# JSON schema validation
jsonschema>=4.17.0

# Rich console output (for better logging)
rich>=13.0.0

# Date/time handling
python-dateutil>=2.8.2

# Google APIs (SEO planning + Sheets/GSC/GA4)
google-api-python-client>=2.130.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-analytics-data>=0.18.6
